// workflow_fe/src/lib/api/gen/types/ResubmitBillingRequestDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ResubmitBillingRequestDto = {
  /**
   * @type string | undefined
   */
  title?: string;
  /**
   * @type string | undefined
   */
  description?: string;
  /**
   * @type string | undefined
   */
  customerName?: string;
  /**
   * @type string | undefined
   */
  customerEmail?: string;
  /**
   * @type string | undefined
   */
  customerAddress?: string;
  /**
   * @minLength 0
   * @type number | undefined
   */
  amount?: number;
  /**
   * @default "BDT"
   * @type string | undefined
   */
  currency?: string;
  /**
   * @type string | undefined
   */
  billingCategory?: string;
  /**
   * @type string | undefined
   */
  departmentId?: string;
  /**
   * @type object | undefined
   */
  customFieldsJson?: object;
};